Refactor DetailsScreen. Create file for BottomSheet
